QuickItemTree: Replace item widgets by a delegate.
In order to be able to animate the text color, the quick item tree had
a widget for each item in the tree view with the text and icons. This
has the consequence of huge widget trees, which results in very bad
performance. Now the same is done using a QItemDelegate.